requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
darkdetect==0.8.0
easyocr==1.6.2
filelock==3.12.0
idna==3.4
imageio==2.28.1
Jinja2==3.1.2
lazy_loader==0.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.3
opencv-python-headless==4.5.4.60
packaging==23.1
Pillow==9.5.0
pyclipper==1.3.0.post4
pyperclip==1.8.2
pytesseract==0.3.10
python-bidi==0.4.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sympy==1.12
tesseract==0.1.3
tifffile==2023.4.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.5.0
urllib3==2.0.2
